+++ internal/infrastructure/controllers/menu_handler.go
@@ -1,4 +1,4 @@
-package http
+package controllers
 
 import (
 	"encoding/json"
